// src/common/types/edit-pengajuan-data.ts
export interface EditPengajuanIzinData {
  keterangan?: string;
  bukti_pendukung?: string;
  tanggal_mulai?: string;
  tanggal_berakhir?: string;
}

export interface EditPengajuanLemburData {
  keterangan?: string;
  bukti_pendukung?: string;
}

export interface EditPengajuanSakitData {
  keterangan?: string;
  bukti_pendukung?: string;
}